fixed, server running in docker now

+++ init.sql
@@ -4,14 +4,14 @@ CREATE DATABASE IF NOT EXISTS appDB CHARACTER SET utf8mb4 COLLATE utf8mb4_unicod
 USE appDB;
 
 CREATE TABLE IF NOT EXISTS category (
-    id BIGINT(11) NOT NULL AUTO_INCREMENT,
+    id INT(11) NOT NULL AUTO_INCREMENT,
     name VARCHAR(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci,
     PRIMARY KEY (ID)
 );
 
 CREATE TABLE IF NOT EXISTS product (
   name VARCHAR(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci,
-  category_id BIGINT NOT NULL,
+  category_id INT NOT NULL,
   foreign key (category_id) references category (id),
   PRIMARY KEY (NAME)
 );
